backend/gemini-api/requirements.txt
requests==2.32.4
fastapi==0.115.14
uvicorn==0.35.0
python-dotenv==1.1.1
google-genai==1.25.0
pydantic==2.11.7
pillow==11.3.0
SQLAlchemy==2.0.41
PyMySQL==1.1.1
python-dotenv==1.1.1
beautifulsoup4==4.13.5
lxml==6.0.1
bleach==6.2.0